% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{tsv_import}
\alias{tsv_import}
\title{Import graph data}
\usage{
tsv_import(tsv_filepath)
}
\arguments{
\item{tsv_filepath}{file path to tsv document generated by copying graph data from pimsoft.}
}
\value{
returns a tsv df, adapted for incorporation to ggplot graph.
}
\description{
Import graph data
}
